# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ArchiveMaster API",
        "description": "Sistema de gestión documental empresarial completo",
        "contact": {
            "name": "ArchiveMaster Support",
            "email": "admin@archivemaster.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "ArchiveMaster API Server"
        }
    ],
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Listar categorías",
                "description": "Obtener lista jerárquica de categorías disponibles para la empresa",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "responses": {
                    "200": {
                        "description": "Lista de categorías",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Contratos"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "contratos"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Documentos contractuales"
                                                    },
                                                    "parent_id": {
                                                        "type": "integer",
                                                        "example": null
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "blue"
                                                    },
                                                    "icon": {
                                                        "type": "string",
                                                        "example": "heroicon-o-document-text"
                                                    },
                                                    "active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "children": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Obtener categoría",
                "description": "Obtener detalles de una categoría específica",
                "operationId": "09767bb2e47dde77dd1e2591428b0ae1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la categoría",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Contratos"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "contratos"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Documentos contractuales"
                                                },
                                                "parent_id": {
                                                    "type": "integer",
                                                    "example": null
                                                },
                                                "color": {
                                                    "type": "string",
                                                    "example": "blue"
                                                },
                                                "icon": {
                                                    "type": "string",
                                                    "example": "heroicon-o-document-text"
                                                },
                                                "active": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "parent": {
                                                    "type": "object"
                                                },
                                                "children": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "documents_count": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usar Bearer Token para autenticación",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}